Implemented request sending feature.
This app now sends one patch request to the server after the user boards a bus.

Known issues:
1. No bus icon appeared on the map after sending a request to the server, while the request and JSON object were in the correct format. Possible causes are
a) testing location was not within 20m of a bus stop (if the server check this as well);
b) testing was not in regular shuttle running time;
c) request or JSON object was still not in the correct format.

Solution:
I will test with a) and b) conditions again and check the HTTP request and JSON object format with other members.

Future to-dos:
1. Check if the user is within 20m of a bus stop.
2. Keep sending a request every 5s with the latest location data if the user is still on the bus.
3. Convert redundant codes into reusable functions.