Create new union schedule view
Shows a bottomsheet peak that prompts user to click to open the new schedule bottom sheet which displays the times for the buses to arrive at the union.

Made it so clicking stops no longer opens the bottomsheet and instead displays the name.

Refactored the buttons on the screen and reorganized functions.